# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Main API interfaces, including chat, session management, authentication, etc..",
        "title": "Morign API",
        "contact": {
            "name": "liut",
            "url": "https://github.com/liut"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/chat": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "聊天"
                ],
                "summary": "发送聊天消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "聊天请求",
                        "name": "chatRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/api.ChatMessage"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求或参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            }
        },
...
        "/api/conversation/{csid}/title": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "聊天"
                ],
                "summary": "生成会话标题",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "csid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            }
        },
...
        "/api/history/{cid}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "聊天"
                ],
                "summary": "获取会话历史",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/aigc.HistoryItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            }
        },
...
        "/api/summary": {
            "post": {
                "description": "根据聊天记录生成简短标题",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "生成聊天记录摘要",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SummaryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Done"
                        }
                    }
                }
            }
        },
        "/api/tools": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "聊天"
                ],
                "summary": "获取可用工具列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/mcps.ToolDescriptor"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "aigc.HistoryChatItem": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "string"
                },
                "th": {
                    "description": "reasoning/thinking content",
                    "type": "string"
                },
                "u": {
                    "type": "string"
                }
            }
        },
        "aigc.HistoryItem": {
            "type": "object",
            "properties": {
                "ci": {
                    "description": "chat",
                    "allOf": [
                        {
                            "$ref": "#/definitions/aigc.HistoryChatItem"
                        }
                    ]
                },
                "ts": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "api.ChatMessage": {
            "type": "object",
            "properties": {
                "csid": {
                    "type": "string"
                },
                "delta": {
                    "type": "string"
                },
                "finishReason": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "think": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tool_call_id": {
                    "description": "For Role=tool prompts this should be set to the ID given in the assistant's prior request to call a tool.",
                    "type": "string"
                },
                "tool_calls": {
                    "description": "For Role=assistant prompts this may be set to the tool calls generated by the model, such as function calls.",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        },
        "api.ChatRequest": {
            "type": "object",
            "properties": {
                "csid": {
                    "type": "string"
                },
                "mcps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "options": {
                    "description": "deprecated: for github.com/Chanzhaoyu/chatgpt-web only",
                    "type": "object",
                    "properties": {
                        "conversationId": {
                            "type": "string"
                        }
                    }
                },
                "pmid": {
                    "description": "TODO: rename to PreviousMessageID",
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "regen": {
                    "type": "boolean"
                },
                "skills": {
                    "description": "技能清单范围：仅圈定注入 system prompt 的技能元数据",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stream": {
                    "type": "boolean"
                }
            }
        },
        "api.Done": {
            "type": "object",
            "properties": {
                "extra": {
                    "description": "附加数据,可选"
                },
                "result": {
                    "description": "主体数据,可选"
                },
                "status": {
                    "description": "状态值，0=ok",
                    "type": "integer",
                    "example": 0
                },
                "t": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
        "api.Failure": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "错误字段,可选,多用于表单校验",
                    "type": "string",
                    "example": "错误字段"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "错误信息"
                },
                "status": {
                    "description": "状态值",
                    "type": "integer",
                    "example": 1
                },
                "t": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
...
        "api.SummaryRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "tips": {
                    "type": "string"
                }
            }
        },
...
        "mcps.ToolDescriptor": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "inputSchema": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "resp.Done": {
            "type": "object",
            "properties": {
                "extra": {
                    "description": "附加数据,可选"
                },
                "result": {
                    "description": "主体数据,可选"
                },
                "status": {
                    "description": "状态值，0=ok",
                    "type": "integer",
                    "example": 0
                },
                "t": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
...
    }
}