# pkg/web/api/handle_convo.go
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"io"
	"net/http"
	"strconv"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/render"
	"github.com/jpillora/eventsource"
	"github.com/marcsv/go-binder/binder"

	"github.com/liut/morign/pkg/models/aigc"
	"github.com/liut/morign/pkg/models/convo"
	"github.com/liut/morign/pkg/models/corpus"
	"github.com/liut/morign/pkg/services/agent"
	"github.com/liut/morign/pkg/services/llm"
	"github.com/liut/morign/pkg/services/stores"
	"github.com/liut/morign/pkg/services/tools"
	"github.com/liut/morign/pkg/settings"
	"github.com/liut/morign/pkg/utils/words"
)

func init() {

	regHI(true, "GET", "/config", "", func(a *api) http.HandlerFunc {
		return a.getConfig
	})
	regHI(true, "GET", "/welcome", "", func(a *api) http.HandlerFunc {
		return a.getWelcome
	})
	regHI(true, "GET", "/history/{cid}", "", func(a *api) http.HandlerFunc {
		return a.getHistory
	})
	regHI(true, "GET", "/tools", "", func(a *api) http.HandlerFunc {
		return a.getTools
	})
	regHI(true, "POST", "/summary", "", func(a *api) http.HandlerFunc {
		return a.postSummary
	})
	regHI(true, "PATCH", "/conversation/{csid}/title", "", func(a *api) http.HandlerFunc {
		return a.patchConversationTitle
	})
}

// chatRequest 内部聊天请求结构
type chatRequest struct {
	messages []llm.Message
	tools    []llm.ToolDefinition
	isSSE    bool
	cs       stores.Conversation
	prompt   string
	skills   []string
}

// prepareSystemMessage 准备系统消息，包括基础 prompt、记忆、工具或知识库
func prepareSystemMessage(ctx context.Context, sto stores.Storage,
	toolreg *tools.Registry, prompt string, skills []string, cs stores.Conversation) (
	llm.Message, []llm.ToolDefinition) {
	var sb strings.Builder

	if len(sto.Preset().SystemPrompt) > 0 {
		sb.WriteString(sto.Preset().SystemPrompt)
	} else {
		sb.WriteString(agent.DefaultSystemMsg)
	}

	if settings.Current.DateInContext {
		sb.WriteString("\n")
		sb.WriteString(agent.ThisMoment())
	}

	fmt.Fprintf(&sb, "\nCurrent SessionID: %s\n", cs.GetID())

	if user, ok := stores.UserFromContext(ctx); ok {
		sb.WriteString("\nCurrent User:\n")
		if user.UID == user.Name {
			fmt.Fprintf(&sb, "Name and uid: %s\n", user.UID)
		} else {
			fmt.Fprintf(&sb, "Name: %s\nuid: %s\n", user.Name, user.UID)
		}
		fmt.Fprintf(&sb, "ID: %s\n", user.OID)

		momories, _, err := sto.Convo().ListMemory(ctx, &stores.ConvoMemorySpec{IsOwner: true})
		if err == nil {
			mtext := momories.PrettyTextForOwner()
			logger().Debugw("load memories", "keys", momories.Keys(), "text", words.TakeTail(mtext, 10, ".."))
			sb.WriteString("\n")
			sb.WriteString(mtext)
		} else {
			logger().Infow("ListMemory fail", "err", err)
		}
	}

	// 转换 MCP 工具为 LLM 工具定义
	tools := agent.ConvertMCPToolsToLLMTools(toolreg.ToolsFor(ctx))
	if len(tools) > 0 {
		toolsPrompt := agent.DefaultToolsMsg
		if len(sto.Preset().ToolsPrompt) > 0 {
			toolsPrompt = sto.Preset().ToolsPrompt
		}
		sb.WriteString("\n")
		sb.WriteString(toolsPrompt)
		cs.SetTools(llm.Tools(tools).Names()...)
	} else {
		docs, err := sto.Corpus().MatchDocments(ctx, stores.MatchSpec{
			Query: prompt,
			Limit: 5,
		})
		if err == nil {
			logger().Infow("matches", "docs", len(docs), "prompt", prompt)
			content := docs.MarkdownText()
			if len(docs) == 0 {
				content += "\nPlease honestly state that you don't know rather than making up an answer."
			}
			sb.WriteString("\n")
			sb.WriteString(content)
		} else {
			logger().Warnw("match fail", "err", err)
			sb.WriteString("\n知识库查询服务暂时不可用，请稍后再试或联系管理员。")
		}
	}

	if len(cs.GetChannel()) > 0 && len(sto.Preset().ChannelPrompt) > 0 {
		sb.WriteString("\n")
		sb.WriteString(sto.Preset().ChannelPrompt)
	}

	appendSkillPrompt(ctx, &sb, sto.Skill(), skills)

	msg := llm.Message{
		Role:    llm.RoleSystem,
		Content: sb.String(),
	}

	return msg, tools
}

// appendSkillPrompt 将技能注入块追加到 system prompt 构建器，失败时降级为空。
func appendSkillPrompt(ctx context.Context, sb *strings.Builder, sk agent.SkillStore, requested []string) {
	block, err := agent.BuildSkillPrompt(ctx, sk, requested)
	if err != nil {
		logger().Warnw("skill prompt fail", "err", err)
		return
	}
	if block != "" {
		sb.WriteString(block)
	}
}

func (a *api) prepareChatRequest(ctx context.Context, param *ChatRequest) *chatRequest {
	cs := stores.NewConversation(ctx, param.GetConversionID())

	sysMsg, tools := prepareSystemMessage(ctx, a.sto, a.toolreg, param.Prompt, param.Skills, cs)
	messages := []llm.Message{sysMsg}

	data, err := cs.ListHistory(ctx)
	if err == nil && len(data) > 0 {
		logger().Infow("found history", "size", len(data), "hist", aigc.HiLogged(data))
		data = data.RecentlyWithTokens(historyLimitToken)

		for i, hi := range data {
			if hi.ChatItem != nil {
				isLast := i == len(data)-1
				isRetry := hi.ChatItem.User == param.Prompt

				// 最后一条特殊处理：如果是重试或 Regenerate，完全跳过这一条历史
				if isLast && (isRetry || param.Regenerate) {
					logger().Debugw("skip last history", "retry", isRetry, "regenerate", param.Regenerate)
					break
				}

				// 添加 User
				if len(hi.ChatItem.User) > 0 {
					messages = append(messages, llm.Message{
						Role: llm.RoleUser, Content: hi.ChatItem.User})
				}

				// 添加 Assistant
				if len(hi.ChatItem.Assistant) > 0 {
					messages = append(messages, llm.Message{
						Role: llm.RoleAssistant, Content: hi.ChatItem.Assistant})
				}
			}
		}
	}

	messages = append(messages, llm.Message{
		Role:    llm.RoleUser,
		Content: param.Prompt,
	})

	return &chatRequest{
		messages: messages,
		tools:    tools,
		cs:       cs,
		prompt:   param.Prompt,
		skills:   param.Skills,
	}
}

// @Tags 聊天
// @Summary 发送聊天消息
// @Accept json
// @Produce json
// @Param token header string false "登录票据凭证"
// @Param chatRequest body ChatRequest true "聊天请求"
// @Success 200 {object} Done{result=ChatMessage}
// @Failure 400 {object} Failure "请求或参数错误"
// @Failure 500 {object} Failure "服务端错误"
// @Router /api/chat [post]
func (a *api) postChat(w http.ResponseWriter, r *http.Request) {
	var param ChatRequest
	if err := binder.BindBody(r, &param); err != nil {
		apiFail(w, r, 400, err)
		return
	}
	isSSE := param.Stream || strings.HasSuffix(r.URL.Path, "-sse")
	isStream := param.Stream || isSSE
	ccr := a.prepareChatRequest(r.Context(), &param)
	ccr.isSSE = isSSE

	loop := agent.NewAgentLoop(agent.AgentLoopConfig{
		LLM:      a.llm,
		ToolExec: a.toolExec,
		MaxLoop:  settings.Current.MaxLoopIterations,
	})

	logger().Infow("chat", "csid", param.GetConversionID(), "msgs", len(ccr.messages), "prompt", param.Prompt, "ip", r.RemoteAddr)

	if isStream {
		a.handleSSEStream(w, r, loop, ccr)
		return
	}

	answer, err := loop.RunNonStreaming(r.Context(), ccr.messages, ccr.tools)
	if err != nil {
		apiFail(w, r, 500, err)
		return
	}
	logger().Infow("chat", "answer", answer)

	var cm ChatMessage
	cm.Text = answer
	render.JSON(w, r, &cm)
}

// writeEvent write and auto flush
func writeEvent(w io.Writer, id string, m any) bool {
	var b []byte
	var err error
	if s, ok := m.(string); ok {
		b = []byte(s)
	} else {
		b, err = json.Marshal(m)
		if err != nil {
			logger().Infow("json marshal fail", "m", m, "err", err)
			return false
		}
	}

	if err = eventsource.WriteEvent(w, eventsource.Event{
		ID:   id,
		Data: b,
	}); err != nil {
		logger().Infow("eventsource write fail", "err", err)
		return false
	}
	if flusher, ok := w.(http.Flusher); ok {
		flusher.Flush()
	}
	return true
}

// handleSSEStream runs the agent loop and writes events to the SSE response.
func (a *api) handleSSEStream(w http.ResponseWriter, r *http.Request, loop *agent.AgentLoop, ccr *chatRequest) {
	if _, ok := w.(http.Flusher); !ok {
		http.Error(w, "Streaming unsupported!", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Cache-Control", "no-cache")
	w.Header().Set("Connection", "keep-alive")
	w.Header().Set("Content-Type", "text/event-stream")
	w.Header().Add("Conversation-ID", ccr.cs.GetID())
	w.(http.Flusher).Flush()

	cctx := stores.ContextWithConvoID(r.Context(), ccr.cs.GetID())

	var (
		answer         string
		think          string
		usage          *llm.Usage
		finish         llm.FinishReason
		chunkIdx       int
		lastWriteEmpty bool
	)

	for event, err := range loop.Run(cctx, ccr.messages, ccr.tools) {
		if err != nil {
			logger().Infow("agent loop error", "err", err)
			break
		}

		if event.ToolResult != nil {
			continue
		}

		var cm ChatMessage
		cm.Delta = event.Delta
		cm.Think = event.Think

		answer += event.Delta
		think += event.Think

		if event.Usage != nil {
			usage = event.Usage
		}

		if event.Done {
			finish = event.StopReason

			if len(event.ToolCalls) > 0 && event.StopReason == llm.FinishReasonToolCalls {
				cm.ToolCalls = convertToolCallsForJSON(event.ToolCalls)
				chunkIdx++
				cm.ConversationID = ccr.cs.GetID()
				cm.FinishReason = string(event.StopReason)
				_ = writeEvent(w, strconv.Itoa(chunkIdx), &cm)
			}

			if event.Usage != nil {
				meta := map[string]any{
					"prompt":     ccr.prompt,
					"answerHead": words.TakeHead(answer, 12, ".."),
					"answerTail": words.TakeTail(answer, 15, ".."),
				}
				if len(event.ResponseID) > 0 {
					meta["resposeID"] = event.ResponseID
				}
				if err := a.rnr.Persist(r.Context(), ccr.cs.GetID(), &llm.Event{
					Author:   "assistant",
					Usage:    event.Usage,
					Model:    event.Model,
					MsgCount: len(ccr.messages),
					Meta:     meta,
				}); err != nil {
					logger().Infow("persist usage fail", "err", err)
				}
			}

			lastWriteEmpty = false
			continue
		}

		isEmpty := event.Delta == "" && event.Think == "" && len(cm.ToolCalls) == 0
		if !isEmpty || !lastWriteEmpty {
			chunkIdx++
			if !writeEvent(w, strconv.Itoa(chunkIdx), &cm) {
				break
			}
		}
		lastWriteEmpty = isEmpty
	}

	logger().Infow("stream response", "answer_len", len(answer), "finish", finish)

	if len(answer) > 0 {
		if err := a.rnr.Persist(r.Context(), ccr.cs.GetID(), &llm.Event{
			Author:     "assistant",
			Delta:      answer,
			Think:      think,
			UserPrompt: ccr.prompt,
		}); err != nil {
			logger().Infow("persist fail", "err", err)
		}
	}

	if settings.Current.QAChatLog && len(answer) > 0 {
		in := corpus.ChatLogBasic{
			ChatID:   ccr.cs.GetOID(),
			Question: ccr.prompt,
			Answer:   answer,
		}
		ip, _, _ := strings.Cut(r.RemoteAddr, ":")
		in.MetaAddKVs("ip", ip)
		if usage != nil {
			in.MetaAddKVs("usage", usage)
		}
		_, err := stores.Sgt().Corpus().CreateChatLog(r.Context(), in)
		if err != nil {
			logger().Infow("save chat log fail", "err", err)
		}
	}

	ctx, cancel := context.WithTimeout(r.Context(), 4*time.Second)
	defer cancel()

	var doneCm ChatMessage
	chunkIdx++
	doneCm.ConversationID = ccr.cs.GetID()
	doneCm.FinishReason = string(finish)

	history, err := ccr.cs.ListHistory(ctx)
	if err == nil && len(history) > 0 {
		title, err := stores.GetHistorySummary(ctx, history)
		if err == nil {
			doneCm.Title = title
		}
	}

	_ = writeEvent(w, strconv.Itoa(chunkIdx), &doneCm)

	chunkIdx++
	_ = writeEvent(w, strconv.Itoa(chunkIdx), esDone)
}

// @Tags 聊天
// @Summary 获取配置信息
// @Accept json
// @Produce json
// @Success 200 {object} Done{result=string}
// @Router /api/config [get]
func (a *api) getConfig(w http.ResponseWriter, r *http.Request) {
	apiOk(w, r, M{
		"apiModel": settings.Current.Interact.Model,
	})
}

// @Tags 聊天
// @Summary 获取欢迎信息
// @Accept json
// @Produce json
// @Success 200 {object} Done{result=string}
// @Router /api/welcome [get]
func (a *api) getWelcome(w http.ResponseWriter, r *http.Request) {
	content := a.preset.Welcome
	if content == "" {
		content = welcomeText
	}

	cs := stores.NewConversation(r.Context(), "")
	apiOk(w, r, M{"id": cs.GetID(), "content": content})
}

// SummaryRequest 摘要请求
type SummaryRequest struct {
	Tips string `json:"tips,omitempty"`
	Text string `json:"text"`
}

// @Summary 生成聊天记录摘要
// @Description 根据聊天记录生成简短标题
// @Accept json
// @Produce json
// @Param request body SummaryRequest true "请求参数"
// @Success 200 {object} resp.Done
// @Router /api/summary [post]
func (a *api) postSummary(w http.ResponseWriter, r *http.Request) {
	var req SummaryRequest
	if err := binder.BindBody(r, &req); err != nil {
		fail(w, r, 400, "invalid request body")
		return
	}
	if req.Text == "" {
		fail(w, r, 400, "text is required")
		return
	}

	summary, err := stores.GetSummary(r.Context(), req.Text, req.Tips)
	if err != nil {
		fail(w, r, 500, err)
		return
	}

	apiOk(w, r, summary)
}

// @Tags 聊天
// @Summary 生成会话标题
// @Accept json
// @Produce json
// @Param token header string false "登录票据凭证"
// @Param csid path string true "会话ID"
// @Success 200 {object} Done{result=string}
// @Failure 400 {object} Failure "请求错误"
// @Failure 500 {object} Failure "服务端错误"
// @Router /api/conversation/{csid}/title [patch]
func (a *api) patchConversationTitle(w http.ResponseWriter, r *http.Request) {
	csid := chi.URLParam(r, "csid")
	if csid == "" {
		fail(w, r, 400, "csid is required")
		return
	}

	// 获取会话历史
	cs := stores.NewConversation(r.Context(), csid)
	history, err := cs.ListHistory(r.Context())
	if err != nil {
		fail(w, r, 500, err)
		return
	}

	if len(history) == 0 {
		fail(w, r, 400, "no history found")
		return
	}

	// 调用 GetHistorySummary 生成标题
	summary, err := stores.GetHistorySummary(r.Context(), history)
	if err != nil {
		fail(w, r, 500, err)
		return
	}

	// 更新会话标题
	title := summary
	err = a.sto.Convo().UpdateSession(r.Context(), csid, convo.SessionSet{Title: &title})
	if err != nil {
		fail(w, r, 500, err)
		return
	}

	apiOk(w, r, M{"title": summary})
}

// @Tags 聊天
// @Summary 获取会话历史
// @Accept json
// @Produce json
// @Param token header string false "登录票据凭证"
// @Param cid path string true "会话ID"
// @Success 200 {object} Done{result=aigc.HistoryItem}
// @Failure 500 {object} Failure "服务端错误"
// @Router /api/history/{cid} [get]
func (a *api) getHistory(w http.ResponseWriter, r *http.Request) {
	cid := chi.URLParam(r, "cid")
	cs := stores.NewConversation(r.Context(), cid)
	data, err := cs.ListHistory(r.Context())
	if err != nil {
		apiFail(w, r, 500, err)
		return
	}
	apiOk(w, r, data, 0)
}

// @Tags 聊天
// @Summary 获取可用工具列表
// @Accept json
// @Produce json
// @Param token header string false "登录票据凭证"
// @Success 200 {object} Done{result=[]mcps.ToolDescriptor}
// @Router /api/tools [get]
func (a *api) getTools(w http.ResponseWriter, r *http.Request) {
	apiOk(w, r, a.toolreg.ToolsFor(r.Context()), 0)
}

// convertToolCallsForJSON 将 llm.ToolCall 转换为可序列化的 map 格式
func convertToolCallsForJSON(tcs []llm.ToolCall) []map[string]any {
	if len(tcs) == 0 {
		return nil
	}
	result := make([]map[string]any, len(tcs))
	for i, tc := range tcs {
		args := string(tc.Function.Arguments)
		if args == "" {
			args = "{}"
		}
		result[i] = map[string]any{
			"id":   tc.ID,
			"type": tc.Type,
			"function": map[string]any{
				"name":      tc.Function.Name,
				"arguments": args,
			},
		}
	}
	return result
}